работает без списка актеров

+++ app/src/main/java/com/example/androidfundamentals2020/MainActivity.kt
@@ -22,7 +22,7 @@ class MainActivity : AppCompatActivity(), MoviesListFragment.OnMoviesListListene
         }
     }
 
-    override fun onMoviesListMovieClicked(movieID: Int) {
+    override fun onMoviesListMovieClicked(movieID: Long) {
         supportFragmentManager.beginTransaction()
             .apply {
                 add(R.id.main_frame_layout, MovieDetailsFragment.newInstance(movieID))
